requirements/base.txt
Django>=5.2,<5.3
psycopg[binary]>=3.1,<4
dj-database-url>=2.1,<3
django-allauth>=65,<66
boto3>=1.34,<2
django-storages[s3]>=1.14,<2
google-api-python-client>=2.100,<3
google-auth-httplib2>=0.2,<1
google-auth-oauthlib>=1.2,<2
django-crispy-forms>=2.1,<3
crispy-tailwind>=1.0,<2
django-htmx>=1.17,<2
Pillow>=10.2,<11
python-decouple>=3.8,<4
